+++ cli/scripts/api.py
@@ -104,41 +104,45 @@ def check_output_wmic(p_cmds):
 
 
 def top(display=True, isJson=False):
-    try:
-        import psutil
-    except ImportError:
-        util.exit_message("Missing psutil module", 1)
+    import pypsutil
 
     current_timestamp = int(time.mktime(datetime.utcnow().timetuple()))
     jsonDict = {}
     procs = []
-    for p in psutil.process_iter():
-        try:
-            p = p.as_dict(
-                attrs=[
-                    "pid",
-                    "username",
-                    "cpu_percent",
-                    "memory_percent",
-                    "cpu_times",
-                    "name",
-                ]
-            )
-        except (psutil.NoSuchProcess, IOError, OSError):
-            pass
-        else:
-            procs.append(p)
+
+    for p in pypsutil.process_iter():
+        print(f"{p.pid: <7}  {p.username(): <10}  {p.memory_percent():.2f}  {p.name()}")
+
+    return
+
+    # TODO:
+        
+    #    try:
+    #        p = p(
+    #            attrs=[
+    #                "pid",
+    #                "username",
+    #                "cpu_percent",
+    #                "memory_percent",
+    #                "cpu_times",
+    #                "name",
+    #            ]
+    #        )
+    #    except (pypsutil.NoSuchProcess, IOError, OSError):
+    #        pass
+    #    else:
+    #        procs.append(p)
 
     if not display:
         return
 
     processes = sorted(procs, key=lambda p: p["cpu_percent"], reverse=True)
 
-    network_usage = psutil.net_io_counters()
+    network_usage = pypsutil.net_io_counters()
     jsonDict["kb_sent"] = network_usage.bytes_sent / 1024
     jsonDict["kb_recv"] = network_usage.bytes_recv / 1024
 
-    cpu = psutil.cpu_times_percent(percpu=False)
+    cpu = pypsutil.cpu_times_percent(percpu=False)
     iowait = ""
     if util.get_platform() == "Linux":
         jsonDict["iowait"] = str(cpu.iowait)
@@ -160,7 +164,7 @@ def top(display=True, isJson=False):
             + iowait
         )
 
-    disk = psutil.disk_io_counters(perdisk=False)
+    disk = pypsutil.disk_io_counters(perdisk=False)
     read_kb = disk.read_bytes / 1024
     write_kb = disk.write_bytes / 1024
     jsonDict["kb_read"] = str(read_kb)
@@ -168,7 +172,7 @@ def top(display=True, isJson=False):
     if not isJson:
         print("DISK: kB_read " + str(read_kb) + ", kB_written " + str(write_kb))
 
-    uptime = datetime.now() - datetime.fromtimestamp(psutil.boot_time())
+    uptime = datetime.now() - datetime.fromtimestamp(pypsutil.boot_time())
     str_uptime = str(uptime).split(".")[0]
     line = ""
     uname_len = 8